requirements.txt
azure-ai-projects==1.1.0b4
azure-ai-agents==1.2.0b3
azure-identity==1.25.1
azure-cli==2.81.0

fastapi==0.123.9
uvicorn==0.38.0

pydantic==2.12.5

requests==2.31.0

fastmcp==2.14.1

python-dotenv==1.2.1
